# Add redact API as first class citizen + faster redaction than pino-noir