--- a/lib/redact.js
+++ b/lib/redact.js
@@ -24,7 +24,13 @@ function redact (opts, serialize) {
   const shape = paths.reduce((o, str) => {
     rx.lastIndex = 0
     rx.exec(str)
-    const { index } = rx.exec(str)
+    const next = rx.exec(str)
+    // top level key:
+    if (next === null) {
+      o[str] = null
+      return o
+    }
+    const { index } = next
     o[str.substr(0, index - 1)] = o[str.substr(0, index - 1)] || []
     o[str.substr(0, index - 1)].push(str.substr(index, str.length - 1))
     return o
@@ -36,9 +42,13 @@ function redact (opts, serialize) {
   const result = {
     [format]: fastRedact({paths, censor, serialize})
   }
+  const serializedCensor = serialize(censor)
+  const topCensor = () => serializedCensor
 
   return Object.keys(shape).reduce((o, k) => {
-    o[k] = fastRedact({paths: shape[k], censor, serialize})
+    // top level key:
+    if (shape[k] === null) o[k] = topCensor
+    else o[k] = fastRedact({paths: shape[k], censor, serialize})
     return o
   }, result)
 }