--- a/lib/redact.js
+++ b/lib/redact.js
@@ -4,20 +4,22 @@ const fastRedact = require('fast-redact')
 const { rx, validator } = fastRedact
 
 const validate = validator({
-  ERR_CENSOR_MUST_BE_FUNCTION: () => 'pino – censor option may not be a function',
-  ERR_PATHS_MUST_BE_STRINGS: () => 'pino – redact option must be an array of strings',
-  ERR_INVALID_PATH: (s) => `pino – redact array contains an invalid path (${s})`
+  ERR_CENSOR_MUST_NOT_BE_FUNCTION: () => 'pino – censor option may not be a function',
+  ERR_PATHS_MUST_BE_STRINGS: () => 'pino – redacted paths must be strings',
+  ERR_INVALID_PATH: (s) => `pino – redact paths array contains an invalid path (${s})`
 })
 
 // symbol for private use
 const format = Symbol('pino.format')
 
+const CENSOR = '[Redacted]'
+
 redact.format = format
 
 module.exports = redact
 
-function redact ({paths, censor, serialize}) {
-  assertValid(paths, censor)
+function redact (opts, serialize) {
+  const { paths, censor } = handle(opts)
 
   const shape = paths.reduce((o, str) => {
     rx.lastIndex = 0
@@ -41,7 +43,16 @@ function redact ({paths, censor, serialize}) {
   }, result)
 }
 
-function assertValid (paths, censor) {
-  if (Array.isArray(paths) === false) { throw Error('pino – redact option must be an array of strings') }
+function handle (opts) {
+  if (Array.isArray(opts)) {
+    opts = {paths: opts, censor: CENSOR}
+    validate(opts)
+    return opts
+  }
+  var { paths, censor = CENSOR, remove } = opts
+  if (Array.isArray(paths) === false) { throw Error('pino – redact must contain an array of strings') }
+  if (remove === true) censor = undefined
   validate({paths, censor})
+
+  return {paths, censor}
 }