--- a/lib/redact.js
+++ b/lib/redact.js
@@ -9,7 +9,14 @@ const validate = validator({
   ERR_INVALID_PATH: (s) => `pino – redact array contains an invalid path (${s})`
 })
 
-module.exports = function redact ({paths, censor, serialize}) {
+// symbol for private use
+const format = Symbol('pino.format')
+
+redact.format = format
+
+module.exports = redact
+
+function redact ({paths, censor, serialize}) {
   assertValid(paths, censor)
 
   const shape = paths.reduce((o, str) => {
@@ -21,10 +28,17 @@ module.exports = function redact ({paths, censor, serialize}) {
     return o
   }, {})
 
+  // the redactor assigned to the format symbol key
+  // provides top level redaction for instances where
+  // an object is interpolated into the msg string
+  const result = {
+    [format]: fastRedact({paths, censor, serialize})
+  }
+
   return Object.keys(shape).reduce((o, k) => {
     o[k] = fastRedact({paths: shape[k], censor, serialize})
     return o
-  }, {})
+  }, result)
 }
 
 function assertValid (paths, censor) {