--- a/lib/redact.js
+++ b/lib/redact.js
@@ -44,7 +44,6 @@ function redact (opts, serialize) {
   }
   const serializedCensor = serialize(censor)
   const topCensor = () => serializedCensor
-
   return Object.keys(shape).reduce((o, k) => {
     // top level key:
     if (shape[k] === null) o[k] = topCensor