@davidmarkclements 2018-05-28
@mcollina happy to add tests and docs - if we're all cool with the API (cc @jsumners )