@mcollina 2018-05-28
I would prefer if we had:

```
require('pino')({
  redact: {
    paths: ['req.headers.cookie'],
    censor: 'hidden :D'
  } // this goes to fast-redact as is.
})
```