@davidmarkclements 2018-05-28
I thought you might say that, we can go that way

It won't go to `fast-redact` as-is though, it has to be processed because of our object name space approach 

if you want the options passed directly you need

```
require('pino')({
  redact: {
    req: {
      paths: ['headers.cookie'],
      censor: 'hidden :D'
    }
  } 
})
```

Which is too verbose for the standard case imo, however it could be supported to allow different censors for different namespaces – this could be useful where on some namespaces you want to remove properties but for others you want to redact