@mcollina 2018-05-28
I'm +1 with the current approach. the other one is too confusing.