@davidmarkclements 2018-05-28
added tests, also updated `quick-format-unescaped` to take a stringifier and released v2 

this solves two problems: 

* keeping the fast-safe-stringify synced, and
* making the format function use the redactor to serialize

There is an implication for browser pino: it will no longer use fast safe stringifier via quick-format-unescaped for interpolated values – instead if will do a try/catch `JSON.stringify` (as per quick-format-unescaped default). We can put this back in, but I'm of the opinion that this is beneficial, it decreases payload size of browser side logging without losing much benefit. We can always put fast-safe-stringify back to `browser.js` instead of course.