@jsumners 2018-05-29
I rather like the array or object config option. It allows flexibility. I'm very onboard with this.