@mcollina 2018-05-29
In the documentation, can you add a caveat section like in https://github.com/davidmarkclements/fast-redact#caveat? We need to note that passing user input in the redact option _is not safe_.