@davidmarkclements 2018-05-29
@mcollina totally agree 
I've updated the api so we have a basic case, `redact: Array` and an advanced case `redact: Object` – this means we drop the top level `censor` option in favour of `redact.censor` and allow for an additional `redact.remove` option

```js
pino({redact: ['a.b.c']})
```

and

```
pino({redact: {
  paths: ['a.b.c'],
  censor: '[Redacted]', //default
  remove: false // default
})
```

Final todo is the docs