@davidmarkclements 2018-05-31
@mcollina Docs have been added

This is no longer a WIP – once we have agreement on approval this can be merged.

🎉